+++ backend/srcs/src/profile_images/profile_images.service.ts
@@ -1,6 +1,5 @@
-import { HttpStatus, Injectable, HttpException } from '@nestjs/common';
+import { Injectable} from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
-import { PrismaClientKnownRequestError } from '@prisma/client/runtime/library';
 import * as fs from 'fs';