+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## 2023-11-09 - carce-bo
+## Added
+ - Added automatic UID / GID to make dev-up etc.
+ - environment generator does not use a tmp file now.
+## changes
+ - Makefile (raiz), generate-environment.sh.
+
 ## 2023-11-09 - carce-bo
 ## Added
 - Profile Images: